flask/flask-template/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements.in
#
blinker==1.6.2
    # via flask
click==8.1.6
    # via flask
flask==2.3.2
    # via -r requirements.in
itsdangerous==2.1.2
    # via flask
jinja2==3.1.2
    # via
    #   -r requirements.in
    #   flask
markupsafe==2.1.3
    # via
    #   jinja2
    #   werkzeug
peewee==3.16.3
    # via -r requirements.in
psycopg2-binary==2.9.7
    # via -r requirements.in
python-dotenv==1.0.0
    # via -r requirements.in
werkzeug==2.3.7
    # via flask
